Support objects (#28)

+++ FakeServer/Controllers/DynamicController.cs
@@ -69,6 +69,30 @@ namespace FakeServer.Controllers
         [HttpHead("{collectionId}")]
         public IActionResult GetItems(string collectionId, int skip = 0, int take = 512)
         {
+            var found = _ds.GetKeys().TryGetValue(collectionId, out var itemType);
+
+            if (found == false)
+                return NotFound();
+
+            if (itemType == JsonFlatFileDataStore.ValueType.Item)
+                return GetSingleItem(collectionId);
+            else
+                return GetCollectionItem(collectionId, skip, take);
+        }
+
+        private IActionResult GetSingleItem(string itemId)
+        {
+            var item = _ds.GetItem(itemId);
+
+            if (item == null)
+                return NotFound();
+
+            // TODO: Add results object or is it needed?
+            return Ok(item);
+        }
+
+        private IActionResult GetCollectionItem(string collectionId, int skip, int take)
+        {
             var collection = _ds.GetCollection(collectionId);
 
             // Collection can actually just be empty, but in this case we handle it as it is not found
@@ -134,11 +158,15 @@ namespace FakeServer.Controllers
         /// <param name="id">Item id</param>
         /// <returns>Item</returns>
         /// <response code="200">Item found</response>
+        /// <response code="400">Item is not collection</response>
         /// <response code="404">Item not found</response>
         [HttpGet("{collectionId}/{id}")]
         [HttpHead("{collectionId}/{id}")]
         public IActionResult GetItem(string collectionId, [FromRoute][DynamicBinder]dynamic id)
         {
+            if (_ds.IsItem(collectionId))
+                return BadRequest();
+
             var result = _ds.GetCollection(collectionId).Find(e => e.id == id).FirstOrDefault();
 
             if (result == null)
@@ -158,12 +186,15 @@ namespace FakeServer.Controllers
         /// <param name="path">Rest of the path</param>
         /// <returns></returns>
         /// <response code="200">Nested item found</response>
-        /// <response code="400">Parent item not found</response>
+        /// <response code="400">Parent item not found or item not collection</response>
         /// <response code="404">Nested item not found</response>
         [HttpGet("{collectionId}/{id}/{*path}")]
         [HttpHead("{collectionId}/{id}/{*path}")]
         public IActionResult GetNested(string collectionId, [FromRoute][DynamicBinder]dynamic id, string path)
         {
+            if (_ds.IsItem(collectionId))
+                return BadRequest();
+
             var item = _ds.GetCollection(collectionId).AsQueryable().FirstOrDefault(e => e.id == id);
 
             if (item == null)
@@ -185,11 +216,15 @@ namespace FakeServer.Controllers
         /// <returns>Created item id</returns>
         /// <response code="201">Item created</response>
         /// <response code="400">Item is null</response>
+        /// <response code="409">Collection is an object</response>
         [HttpPost("{collectionId}")]
         public async Task<IActionResult> AddNewItem(string collectionId, [FromBody]JToken item)
         {
             if (item == null)
                 return BadRequest();
+
+            if (_ds.IsItem(collectionId))
+                return Conflict();
 
             var collection = _ds.GetCollection(collectionId);
 
@@ -206,12 +241,12 @@ namespace FakeServer.Controllers
         /// <param name="item">Item's new content</param>
         /// <returns></returns>
         /// <response code="204">Item found and replaced</response>
-        /// <response code="400">Item is null</response>
+        /// <response code="400">Replace data is null or item is not in a collection</response>
         /// <response code="404">Item not found</response>
         [HttpPut("{collectionId}/{id}")]
         public async Task<IActionResult> ReplaceItem(string collectionId, [FromRoute][DynamicBinder]dynamic id, [FromBody]dynamic item)
         {
-            if (item == null)
+            if (item == null || _ds.IsItem(collectionId))
                 return BadRequest();
 
             // Make sure that new data has id field correctly
@@ -237,19 +272,19 @@ namespace FakeServer.Controllers
         ///        "boolField": true
         ///     }
         /// </remarks>
-        /// <param name="collectionId"></param>
+        /// <param name="collectionId">Collection id</param>
         /// <param name="id">Id of the item to be updated</param>
         /// <param name="patchData">Patch data</param>
         /// <returns></returns>
         /// <response code="204">Item found and updated</response>
-        /// <response code="400">Patch data is empty</response>
+        /// <response code="400">Patch data is empty or item is not in a collection</response>
         /// <response code="404">Item not found</response>
         [HttpPatch("{collectionId}/{id}")]
         public async Task<IActionResult> UpdateItem(string collectionId, [FromRoute][DynamicBinder]dynamic id, [FromBody]JToken patchData)
         {
             dynamic sourceData = JsonConvert.DeserializeObject<ExpandoObject>(patchData.ToString());
 
-            if (!((IDictionary<string, object>)sourceData).Any())
+            if (!((IDictionary<string, object>)sourceData).Any() || _ds.IsItem(collectionId))
                 return BadRequest();
 
             var success = await _ds.GetCollection(collectionId).UpdateOneAsync((Predicate<dynamic>)(e => e.id == id), sourceData);
@@ -267,11 +302,97 @@ namespace FakeServer.Controllers
         /// <param name="id">Id of the item to be removed</param>
         /// <returns></returns>
         /// <response code="204">Item found and removed</response>
+        /// <response code="400">Item is not in a collection</response>
         /// <response code="404">Item not found</response>
         [HttpDelete("{collectionId}/{id}")]
         public async Task<IActionResult> DeleteItem(string collectionId, [FromRoute][DynamicBinder]dynamic id)
         {
+            if (_ds.IsItem(collectionId))
+                return BadRequest();
+
             var success = await _ds.GetCollection(collectionId).DeleteOneAsync(e => e.id == id);
+
+            if (success)
+                return NoContent();
+            else
+                return NotFound();
+        }
+
+        /// <summary>
+        /// Replace object
+        /// </summary>
+        /// <param name="objectId">Object id</param>
+        /// <param name="item">Object's new content</param>
+        /// <returns></returns>
+        /// <response code="204">Object found and replaced</response>
+        /// <response code="400">Replace data is null or item is in a collection</response>
+        /// <response code="404">Object not found</response>
+        [HttpPut("{objectId}")]
+        public async Task<IActionResult> ReplaceSingleItem(string objectId, [FromBody]dynamic item)
+        {
+            if (_ds.IsCollection(objectId))
+                return BadRequest();
+
+            if (item == null)
+                return BadRequest();
+
+            var success = await _ds.ReplaceItemAsync(objectId, item, _settings.UpsertOnPut);
+
+            if (success)
+                return NoContent();
+            else
+                return NotFound();
+        }
+
+        /// <summary>
+        /// Update single object's content
+        /// </summary>
+        /// <remarks>
+        /// Patch data contains fields to be updated.
+        ///
+        ///     {
+        ///        "stringField": "some value",
+        ///        "intField": 22,
+        ///        "boolField": true
+        ///     }
+        /// </remarks>
+        /// <param name="objectId">Object id</param>
+        /// <param name="patchData">Patch data</param>
+        /// <returns></returns>
+        /// <response code="204">Object found and updated</response>
+        /// <response code="400">Patch data is empty</response>
+        /// <response code="404">Object not found</response>
+        [HttpPatch("{objectId}")]
+        public async Task<IActionResult> UpdateSingleItem(string objectId, [FromBody]JToken patchData)
+        {
+            dynamic sourceData = JsonConvert.DeserializeObject<ExpandoObject>(patchData.ToString());
+
+            if (!((IDictionary<string, object>)sourceData).Any() || _ds.IsCollection(objectId))
+                return BadRequest();
+
+            var success = await _ds.UpdateItemAsync(objectId, sourceData);
+
+            if (success)
+                return NoContent();
+            else
+                return NotFound();
+        }
+
+        /// <summary>
+        /// Remove single object
+        /// </summary>
+        /// <param name="objectId">Single object id</param>
+        /// <returns></returns>
+        /// <response code="204">Object found and removed</response>
+        /// <response code="400">Object is a collection</response>
+        /// <response code="404">Object not found</response>
+        [HttpDelete("{objectId}")]
+        public async Task<IActionResult> DeleteSingleItem(string objectId)
+        {
+            if (_ds.IsCollection(objectId))
+                return BadRequest();
+
+            var success = await _ds.DeleteItemAsync(objectId);
 
             if (success)
                 return NoContent();
